feat: separate responsibilities
