--- modulo05/src/routes/index.ts
@@ -1,14 +1,0 @@
-import { Router } from 'express';
-
-import appointmentsRouter from './appointments.routes';
-import usersRouter from './users.routes';
-import sessionsRouter from './sessions.routes';
-
-const routes = Router();
-
-routes.use('/sessions', sessionsRouter);
-
-routes.use('/appointments', appointmentsRouter);
-routes.use('/users', usersRouter);
-
-export default routes;
