+++ modulo05/src/shared/errors/AppError.ts
