+++ modulo05/src/shared/infra/http/middlewares/HandleException.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 
-import AppError from '../errors/AppError';
+import AppError from '@shared/errors/AppError';
 
 export default (
     error: Error,
