+++ modulo05/src/shared/infra/http/routes/index.ts
@@ -1,0 +1,14 @@
+import { Router } from 'express';
+
+import appointmentsRouter from '@modules/appointments/infra/http/routes/appointments.routes';
+import usersRouter from '@modules/users/infra/http/routes/users.routes';
+import sessionsRouter from '@modules/users/infra/http/routes/sessions.routes';
+
+const routes = Router();
+
+routes.use('/sessions', sessionsRouter);
+
+routes.use('/appointments', appointmentsRouter);
+routes.use('/users', usersRouter);
+
+export default routes;
